EBR-121: add .version file for keeping generated folders in sync

+++ tvbextxircuits/utils.py
@@ -1,3 +1,5 @@
+from pathlib import Path
+
 STORAGE_CONFIG_FILE = 'storage_config.json'  # To be used only for HPC runs
 COLLAB_NAME_KEY = 'collab_name'  # Used only for HPC runs
 BUCKET_NAME_KEY = 'bucket_name'  # Used only for HPC runs
@@ -21,13 +23,17 @@ def is_valid_url(url):
     except ValueError:
         return False
 
-def copy_from_installed_wheel(package_name, resource="", dest_path=None):
+def copy_from_installed_wheel(package_name, resource="", dest_path=None, version_changed=False):
     if dest_path is None:
         dest_path = package_name
 
     # Get the resource reference
     ref = importlib_resources.files(package_name) / resource
 
+    config_path = Path(os.getcwd()) / dest_path
     # Create the temporary file context
     with importlib_resources.as_file(ref) as resource_path:
-        shutil.copytree(resource_path, dest_path)
+        if not config_path.exists() or version_changed:
+            if config_path.exists():
+                shutil.rmtree(config_path)
+            shutil.copytree(resource_path, dest_path)
